Note saving error fixes

+++ models/users/user.py
@@ -196,7 +196,7 @@ class User(object):
 
     def delete_user_boxes(self):
         boxes = Box.find_by_maker_id(self._id)
-
-        for box in boxes:
-            box.delete_on_elastic()
-            box.delete()
+        if boxes is not None:
+            for box in boxes:
+                box.delete_on_elastic()
+                box.delete()
